# html/web/openapi.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: AccuMed OpenAPI
  contact:
    email: info@accumedintel.com
...
paths:
...
  /organization/search:
    get:
      tags:
      - Organization
      summary: Search organizations
      operationId: searchOrganization
      parameters:
      - name: search
        in: query
        description: Search string
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  Organizations:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Organization'
                        - $ref: '#/components/schemas/Link'
                  _links:
                    $ref: '#/components/schemas/PageLink'
                  _meta:
                    $ref: '#/components/schemas/Meta'
  /organization/{id}:
    get:
      tags:
      - Organization
      summary: Get organization by ID
      operationId: getOrganizationByID
      parameters:
      - name: id
        in: path
        description: ID of organization
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Organization'
                  - $ref: '#/components/schemas/Link'
    delete:
      tags:
      - Organization
      summary: Delete organization by ID
      operationId: deleteOrganizationByID
      parameters:
      - name: id
        in: path
        description: ID of organization
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content: {}
...
  /user/search:
    get:
      tags:
      - User
      summary: Search Users
      operationId: searchUser
      parameters:
      - name: search
        in: query
        description: Search string
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  Users:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/User'
                        - $ref: '#/components/schemas/Link'
                  _links:
                    $ref: '#/components/schemas/PageLink'
                  _meta:
                    $ref: '#/components/schemas/Meta'
  /user/{id}:
    get:
      tags:
      - User
      summary: Get user by ID
      operationId: getUserByID
      parameters:
      - name: id
        in: path
        description: ID of user
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/Link'
    delete:
      tags:
      - User
      summary: Delete user by ID
      operationId: deleteUserByID
      parameters:
      - name: id
        in: path
        description: ID of user
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
          content: {}
...
components:
  schemas:
...
    Organization:
      type: object
      properties:
        organizationID:
          type: string
        organizationGUID:
          type: string
        name:
          type: string
        fullAddress:
          type: string
        address:
          type: string
        subPremises:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
    User:
      type: object
      properties:
        userID:
          type: string
        userGUID:
          type: string
        name:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        email:
          type: string
        phone:
          type: string
        directLine:
          type: string
        directExtension:
          type: string
        cellphone:
          type: string
        other:
          type: string
...
    Link:
      type: object
      properties:
        _links:
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  type: string
    PageLink:
      type: object
      properties:
        self:
          type: object
          properties:
            href:
              type: string
        first:
          type: object
          properties:
            href:
              type: string
        prev:
          type: object
          properties:
            href:
              type: string
        next:
          type: object
          properties:
            href:
              type: string
        last:
          type: object
          properties:
            href:
              type: string
    Meta:
      type: object
      properties:
        totalCount:
          type: integer
        pageCount:
          type: integer
        currentPage:
          type: integer
        perPage:
          type: integer